Another major cleanup of the castor code targetting to drop all warnings due to the -Wextra flag. On the way, many little bugs were fixed (possible core dumps in command lines, wrong or missing checks, ...), code was dropped since unused, files were converted from dos to unix end of line convention, etc, etc ...

+++ client/src/stager/stager_errmsg.c
@@ -161,7 +161,7 @@ int DLL_DECL stager_errmsg(const char *func, const char *msg, ...) {
 		logfunction(STAGER_MSG_ERR, prtbuf);
 	} else {
 		if (errbufp != NULL && errbuflen > 0) {
-			if (strlen (prtbuf) < errbuflen) {
+			if ((int)strlen (prtbuf) < errbuflen) {
 				strcpy (errbufp, prtbuf);
 			} else {
 				strncpy (errbufp, prtbuf, errbuflen - 1);
@@ -212,7 +212,7 @@ int DLL_DECL stager_outmsg(const char *func, const char *msg, ...) {
 		logfunction(STAGER_MSG_OUT, prtbuf);
 	} else {
 		if (outbufp != NULL && outbuflen > 0) {
-			if (strlen (prtbuf) < outbuflen) {
+			if ((int)strlen (prtbuf) < outbuflen) {
 				strcpy (outbufp, prtbuf);
 			} else {
 				strncpy (outbufp, prtbuf, outbuflen - 1);
