+++ rfio/fchown.c
@@ -79,7 +79,7 @@ int DLL_DECL rfio_fchown(s, owner, group)
     LONG msgsiz ;
 
     TRACE(2, "rfio", "rfio_fchown: reading %d bytes",rfilefdt[s_index]->_iobuf.hsize) ;
-    if (netread_timeout(s,rfio_buf,rfilefdt[s_index]->_iobuf.hsize,RFIO_DATA_TIMEOUT) != rfilefdt[s_index]->_iobuf.hsize) {
+    if (netread_timeout(s,rfio_buf,rfilefdt[s_index]->_iobuf.hsize,RFIO_DATA_TIMEOUT) != (ssize_t)rfilefdt[s_index]->_iobuf.hsize) {
       TRACE(2, "rfio", "rfio_fchown: read(): ERROR occured (errno=%d)", errno);
       if ( temp ) (void) free(trp) ;
       END_TRACE() ;
@@ -105,7 +105,7 @@ int DLL_DECL rfio_fchown(s, owner, group)
        * to receive data which is going to be thrown away.
        */
       if ( temp == 0 ) {
-        if ( rfilefdt[s_index]->_iobuf.base==NULL || rfilefdt[s_index]->_iobuf.dsize<msgsiz ) {
+        if ( rfilefdt[s_index]->_iobuf.base==NULL || (int)rfilefdt[s_index]->_iobuf.dsize<msgsiz ) {
           temp= 1 ;
           TRACE(3,"rfio","rfio_fchown: allocating momentary buffer of size %d",msgsiz) ;
           if ( (trp= ( char *) malloc(msgsiz)) == NULL ) {
